+++ src/models/_entities/shipping_templates.rs
@@ -10,8 +10,8 @@ pub struct Model {
     pub id: i32,
     pub name: String,
     #[sea_orm(column_name = "type")]
-    pub r#type: i32,
-    pub appoint: i32,
+    pub r#type: i16,
+    pub appoint: i16,
     pub sort: i32,
     pub create_time: DateTime,
     pub update_time: DateTime,
