+++ src/models/_entities/shipping_templates_region.rs
@@ -17,9 +17,9 @@ pub struct Model {
     pub renewal: Decimal,
     pub renewal_price: Decimal,
     #[sea_orm(column_name = "type")]
-    pub r#type: i32,
+    pub r#type: i16,
     pub uniqid: String,
-    pub status: Option<i32>,
+    pub status: Option<i16>,
     pub create_time: DateTime,
     pub update_time: DateTime,
 }
